Added finished screen state to avoid duplicate events.

+++ app/src/main/java/com/javisc/roomexample/datasource/service/PhotoApi.kt
@@ -5,9 +5,9 @@ import arrow.core.Try
 import retrofit2.await
 
 interface PhotoApi {
-    suspend fun getPhoto(id: Long): Either<Throwable, PhotoDto>
+    suspend fun getPhoto(id: Int): Either<Throwable, PhotoDto>
 }
 
 class PhotoApiImpl(private val photoService: PhotoService) : PhotoApi {
-    override suspend fun getPhoto(id: Long) = Try { photoService.getPhoto(id).await() }.toEither()
+    override suspend fun getPhoto(id: Int) = Try { photoService.getPhoto(id).await() }.toEither()
 }
